+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added grammar support for eating events with behavior text `MyEvent/`.
   - Before this change, you could use behavior text `MyEvent` instead.
   - https://github.com/StateSmith/StateSmith/issues/43
+- Added validation to prevent `enter`/`exit` triggers on transition behaviors.
+  - https://github.com/StateSmith/StateSmith/issues/162
 
 ---